requirements.txt
blinker==1.9.0
click==8.1.8
dnspython==2.7.0
email-validator==2.2.0
Flask==3.1.0
Flask-Login==0.6.3
flask-mail==0.10.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
python-dotenv==1.1.0
SQLAlchemy==2.0.40
typing-extensions==4.13.2
Werkzeug==3.1.3
WTForms==3.2.1
flask_limiter==3.11.0
gunicorn==23.0.0
streamlit==1.44.1
celery==5.5.2
openai==1.78.0
